SAD-DT_Project/requirements.txt

torch
torchvision
timm
opencv-python
numpy
matplotlib
Pillow
